add oauth response under continueMiddleware
Adds the oauth server response to the res.locals.oauth object when the continueMiddleware option is used. This will allow simple overrides that require continueMiddleware to still have access to the response that was worked on, allowing the developer to use functions similar to the handleResponse to quickly build the express response.